ttd_v2/ttd_v2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "truth-through-debate"
version = "2.0.0"
description = "Multi-agent LLM debate framework for fact-checking"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["llm", "fact-checking", "multi-agent", "debate", "hallucination"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "groq>=0.9.0",
    "aiohttp>=3.9.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.7.0",
    "rank-bm25>=0.2.2",
    "datasets>=2.20.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "gradio>=4.36.0",
    "tavily-python>=0.3.0",
    "pyarrow>=15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
ttd = "truth_through_debate.cli:main"
ttd-ui = "truth_through_debate.ui.app:launch"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/truth-through-debate"
Repository = "https://github.com/YOUR_USERNAME/truth-through-debate"
"Bug Tracker" = "https://github.com/YOUR_USERNAME/truth-through-debate/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["truth_through_debate*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
